fix(l10n): #2148 Move getLocalizedStrings to reducers
