+++ addon/Feeds/LocalizationFeed.js
@@ -2,19 +2,8 @@ const {PrefsTarget} = require("sdk/preferences/event-target");
 const {findClosestLocale, getPreferedLocales} = require("sdk/l10n/locale");
 const Feed = require("../lib/Feed");
 const am = require("../../common/action-manager");
-const DEFAULT_LOCALE = "en-US";
-const STRINGS = require("../../data/locales/locales.json");
-const AVAILABLE_LOCALES = Object.keys(STRINGS);
+const AVAILABLE_LOCALES = Object.keys(require("../../data/locales/locales.json"));
 const RTL_LIST = ["ar", "he", "fa", "ur"];
-
-function getLocalizedStrings(locale, allStrings = STRINGS) {
-  if (locale === DEFAULT_LOCALE) {
-    return allStrings[DEFAULT_LOCALE];
-  }
-  const strings = allStrings[locale];
-  // This will include the English string for any missing ids
-  return Object.assign({}, allStrings[DEFAULT_LOCALE], strings || {});
-}
 
 function getDirection(locale) {
   return (RTL_LIST.indexOf(locale.split("-")[0]) >= 0) ? "rtl" : "ltr";
@@ -45,9 +34,8 @@ class LocalizationFeed extends Feed {
   }
   getData() {
     let locale = findClosestLocale(this.availableLocales, getPreferedLocales());
-    const strings = getLocalizedStrings(locale);
     const direction = getDirection(locale);
-    return Promise.resolve(am.actions.Response("LOCALE_UPDATED", {locale, strings, direction}));
+    return Promise.resolve(am.actions.Response("LOCALE_UPDATED", {locale, direction}));
   }
   onAction(state, action) {
     switch (action.type) {
@@ -61,10 +49,7 @@ class LocalizationFeed extends Feed {
   }
 }
 
-LocalizationFeed.DEFAULT_LOCALE = DEFAULT_LOCALE;
-LocalizationFeed.AVAILABLE_LOCALES = AVAILABLE_LOCALES;
 LocalizationFeed.LOCALE_PREFS = LOCALE_PREFS;
-LocalizationFeed.getLocalizedStrings = getLocalizedStrings;
 LocalizationFeed.getDirection = getDirection;
 
 module.exports = LocalizationFeed;
