+++ common/getLocalizedStrings.js
@@ -1,0 +1,11 @@
+const DEFAULT_LOCALE = "en-US";
+const STRINGS = require("../data/locales/locales.json");
+
+module.exports = function getLocalizedStrings(locale, allStrings = STRINGS) {
+  if (locale === DEFAULT_LOCALE) {
+    return allStrings[DEFAULT_LOCALE];
+  }
+  const strings = allStrings[locale];
+  // This will include the English string for any missing ids
+  return Object.assign({}, allStrings[DEFAULT_LOCALE], strings || {});
+};
